Moved functionality provided by each file into its own package. Additionally, TextArea, and TagItem were capitalized so that they can be exported and were added to there own files to ensure we don't have circular imports
