+++ inject/tagitem.go
@@ -1,0 +1,63 @@
+package inject
+
+import (
+	"fmt"
+	"strings"
+)
+
+func TagFromComment(comment string) (tag string) {
+	match := CommentRegex.FindStringSubmatch(comment)
+	if len(match) == 2 {
+		tag = match[1]
+	}
+	return
+}
+
+type TagItem struct {
+	key   string
+	value string
+}
+
+type tagItems []TagItem
+
+func (ti tagItems) format() string {
+	tags := []string{}
+	for _, item := range ti {
+		tags = append(tags, fmt.Sprintf(`%s:%s`, item.key, item.value))
+	}
+	return strings.Join(tags, " ")
+}
+
+func (ti tagItems) override(nti tagItems) tagItems {
+	overrided := []TagItem{}
+	for i := range ti {
+		dup := -1
+		for j := range nti {
+			if ti[i].key == nti[j].key {
+				dup = j
+				break
+			}
+		}
+		if dup == -1 {
+			overrided = append(overrided, ti[i])
+		} else {
+			overrided = append(overrided, nti[dup])
+			nti = append(nti[:dup], nti[dup+1:]...)
+		}
+	}
+	return append(overrided, nti...)
+}
+
+func newTagItems(tag string) tagItems {
+	items := []TagItem{}
+	splitted := TagsRegex.FindAllString(tag, -1)
+
+	for _, t := range splitted {
+		sepPos := strings.Index(t, ":")
+		items = append(items, TagItem{
+			key:   t[:sepPos],
+			value: t[sepPos+1:],
+		})
+	}
+	return items
+}
